+++ lib/src/combinations.dart
@@ -1,36 +1,52 @@
-part of trotter;
+import "combinatorics.dart" show Combinatorics;
+import "helpers.dart"
+    show
+        nCr,
+        itemsAreUnique,
+        indexFromIntOrBigInt,
+        combination,
+        adjustedIndex,
+        inverseCombination,
+        itemsExistInUniversal;
 
 /// A pseudo-list of combinations.
 ///
-/// A pseudo-list 'containing' all the `r`-combinations of objects taken from
+/// A pseudo-list "containing" all the `r`-combinations of objects taken from
 /// the list `items`.
 ///
 /// _Example_
 ///
-///     final c = Combinations(3, characters('abcd'));
-///     print('There are ${c.length} 3-combinations of the objects');
-///     print('in ${c.items}.');
-///     print('The first combination is ${c[0]}.');
+///     final c = Combinations(3, characters("abcd"));
+///     print("There are ${c.length} 3-combinations of the objects");
+///     print("in ${c.items}.");
+///     print("The first combination is ${c[0]}.");
 ///
 
 class Combinations<T> extends Combinatorics<T> {
-  Combinations(this.r, List<T> items) {
+  Combinations(this.r, List<T> items)
+      : _items = List<T>.from(items, growable: false),
+        _length = nCr(items.length, r) {
     if (r < 0 || r > items.length) {
-      throw Exception('Cannot take $r items from ${items.length}.');
+      throw Exception("Cannot take $r items from ${items.length}.");
     }
-    if (!_itemsAreUnique(items)) throw Exception('Items are not unique.');
-
-    _items = List<T>.from(items, growable: false);
-    _length = _nCr(items.length, r);
+    if (!itemsAreUnique(items)) throw Exception("Items are not unique.");
   }
 
   /// The number of items taken from `items`.
   final int r;
 
+  final List<T> _items;
+  @override
+  List<T> get items => List<T>.from(_items, growable: false);
+
+  final BigInt _length;
+  @override
+  BigInt get length => _length;
+
   @override
   List<T> operator [](Object k) {
-    BigInt biK = _indexFromIntOrBigInt(k);
-    return _combination(_adjustedIndex(biK, length), r, items);
+    BigInt biK = indexFromIntOrBigInt(k);
+    return combination(adjustedIndex(biK, length), r, items);
   }
 
   /// Returns the index of `combination` in the list of
@@ -39,7 +55,7 @@ class Combinations<T> extends Combinatorics<T> {
   BigInt indexOf(List<T> combination, [BigInt? start]) {
     start = start ?? BigInt.zero;
     if (contains(combination)) {
-      BigInt result = _inverseCombination(combination, _items);
+      BigInt result = inverseCombination(combination, _items);
       if (result >= start) {
         return result;
       } else {
@@ -53,9 +69,9 @@ class Combinations<T> extends Combinatorics<T> {
   /// Returns whether `x` is in the pseudo-list.
   @override
   bool contains(List<T> x) =>
-      _itemsExistInUniversal(x, _items) && _itemsAreUnique(x);
+      itemsExistInUniversal(x, _items) && itemsAreUnique(x);
 
   @override
   String toString() =>
-      'Pseudo-list containing all $length $r-combinations of items from $items.';
+      "Pseudo-list containing all $length $r-combinations of items from $items.";
 }