+++ lib/src/combinatorics.dart
@@ -1,33 +1,37 @@
-part of trotter;
+import "dart:math" show Random;
+import "helpers.dart" show adjustedIndex, indexFromIntOrBigInt;
 
-/// The abstract parent to the other classes defined in this library.
+/// The base class for the other classes in this library.
 abstract class Combinatorics<T> {
-  late final List<T> _items;
-  late final BigInt _length;
+  //late final List<T> _items;
+  //late final BigInt _length;
 
   /// The list from which the objects are selected
-  List<T> get items => List<T>.from(_items, growable: false);
+  List<T> get items; // => List<T>.from(_items, growable: false);
+
+  /// The number of arrangements "contained" in this pseudo-list.
+  BigInt get length => BigInt.zero;
 
   Iterable<List<T>> _range(BigInt from, BigInt to) sync* {
     do {
-      yield this[_adjustedIndex(from, length)];
+      yield this[adjustedIndex(from, length)];
       from += BigInt.one;
-    } while (_adjustedIndex(from, length) != _adjustedIndex(to, length));
+    } while (adjustedIndex(from, length) != adjustedIndex(to, length));
   }
 
   /// Returns a range of arrangements.
   ///
-  /// The arrangements 'stored' in this pseudo-list from index `from`
+  /// The arrangements "stored" in this pseudo-list from index `from`
   /// up to but not including `to`.
   ///
   Iterable<List<T>> range(Object fromTo, [Object? to]) {
-    BigInt biFrom = _indexFromIntOrBigInt(fromTo), biTo;
+    BigInt biFrom = indexFromIntOrBigInt(fromTo), biTo;
 
     if (to == null) {
       biTo = biFrom;
       biFrom = BigInt.zero;
     } else {
-      biTo = _indexFromIntOrBigInt(to);
+      biTo = indexFromIntOrBigInt(to);
     }
 
     if (biFrom == biTo) {
@@ -48,7 +52,7 @@ abstract class Combinatorics<T> {
       } else if (fromTo is BigInt) {
         biFrom = fromTo;
       } else {
-        throw Exception('Expecting int or BigInt in range.');
+        throw Exception("Expecting int or BigInt in range.");
       }
       if (to == null) {
         biTo = biFrom;
@@ -59,7 +63,7 @@ abstract class Combinatorics<T> {
         } else if (to is BigInt) {
           biTo = to;
         } else {
-          throw Exception('Expecting int or BigInt in range.');
+          throw Exception("Expecting int or BigInt in range.");
         }
       }
     }
@@ -67,11 +71,11 @@ abstract class Combinatorics<T> {
     return _range(biFrom, biTo);
   }
 
-  /// An `Iterable` object that 'contains' all the arrangements.
+  /// An `Iterable` object that "contains" all the arrangements.
   ///
   /// Example:
   ///
-  ///     final permutations = Permutations(3, characters('abcd'));
+  ///     final permutations = Permutations(3, characters("abcd"));
   ///     for (final permutation in permutations.iterable) {
   ///       print(permutation);
   ///     }
@@ -79,27 +83,25 @@ abstract class Combinatorics<T> {
   /// The `iterable` property can also be obtained by simply calling the
   /// combinatoric object. For example, the above code is equivalent to:
   ///
-  ///     final permutations = Permutations(3, characters('abcd'));
+  ///     final permutations = Permutations(3, characters("abcd"));
   ///     for (final permutation in permutations()) {
   ///       print(permutation);
   ///     }
   ///
   Iterable get iterable => this();
 
-  /// The number of arrangements 'contained' in this pseudo-list.
-  BigInt get length => _length;
-
   /// Generates a random sample of arrangements from this pseudo-list.
   Iterable sample(int n, {int? seed, bool withReplacement = false}) {
+    final length = this.length;
     BigInt indexGenerator() {
       final rand = seed == null ? Random() : Random(seed);
-      var index = BigInt.zero, bits = _length.bitLength + 1;
+      var index = BigInt.zero, bits = length.bitLength + 1;
       for (var i = 0; i < bits; i++) {
         index += rand.nextBool() ? BigInt.one : BigInt.zero;
         if (i < bits - 1) {
           index <<= 1;
         }
-        if (index > _length - BigInt.one) {
+        if (index > length - BigInt.one) {
           index = BigInt.zero;
           i = 0;
         }
@@ -111,8 +113,8 @@ abstract class Combinatorics<T> {
       return List<BigInt>.generate(n, (_) => indexGenerator())
           .map((i) => this[i]);
     } else {
-      if (BigInt.from(n) > _length) {
-        throw Exception("Cannot take more than $_length without replacement.");
+      if (BigInt.from(n) > length) {
+        throw Exception("Cannot take more than $length without replacement.");
       }
       final indices = <BigInt>{};
       while (indices.length < n) {
@@ -123,10 +125,12 @@ abstract class Combinatorics<T> {
   }
 
   /// The `k`th arrangement.
-  List<T> operator [](Object k);
+  List<T> operator [](Object k) {
+    throw Error();
+  }
 
   /// The index of  `arrangement`.
-  BigInt indexOf(List<T> arrangement, [BigInt start]);
+  BigInt indexOf(List<T> arrangement, [BigInt? start]);
 
   /// Whether the structure contains `arrangement`.
   bool contains(List<T> arrangement);