+++ lib/src/compositions.dart
@@ -1,35 +1,50 @@
-part of trotter;
+import "combinatorics.dart" show Combinatorics;
+import "helpers.dart"
+    show
+        nCr,
+        itemsAreUnique,
+        indexFromIntOrBigInt,
+        composition,
+        adjustedIndex,
+        inverseComposition,
+        itemsExistInUniversal;
 
 /// A pseudo-list of Compositions (combinations with repetition).
 ///
-/// A pseudo-list 'containing' all the `r`-Compositions of objects taken from
+/// A pseudo-list "containing" all the `r`-Compositions of objects taken from
 /// the list `items`.
 ///
 /// _Example_
 ///
-///     final s = Compositions(3, characters('abcd'));
-///     print('There are ${s.length} 3-Compositions of the objects');
-///     print('in ${s.items}.');
-///     print('The first selection is ${c[0]}.');
+///     final s = Compositions(3, characters("abcd"));
+///     print("There are ${s.length} 3-Compositions of the objects");
+///     print("in ${s.items}.");
+///     print("The first selection is ${c[0]}.");
 ///
 
 class Compositions<T> extends Combinatorics<T> {
-  Compositions(this.r, List<T> items) {
-    if (r < 0) throw Exception('Cannot take $r items from ${items.length}.');
-    if (!_itemsAreUnique(items)) throw Exception('Items are not unique.');
-
-    _items = List<T>.from(items);
-
-    _length = _nCr(items.length + r - 1, r);
+  Compositions(this.r, List<T> items)
+      : _items = List<T>.from(items, growable: false),
+        _length = nCr(items.length + r - 1, r) {
+    if (r < 0) throw Exception("Cannot take $r items from ${items.length}.");
+    if (!itemsAreUnique(items)) throw Exception("Items are not unique.");
   }
 
   /// The number of items taken from `items`.
   final int r;
 
+  final List<T> _items;
+  @override
+  List<T> get items => List<T>.from(_items, growable: false);
+
+  final BigInt _length;
+  @override
+  BigInt get length => _length;
+
   @override
   List<T> operator [](Object k) {
-    BigInt biK = _indexFromIntOrBigInt(k);
-    return _composition(_adjustedIndex(biK, length), r, items);
+    BigInt biK = indexFromIntOrBigInt(k);
+    return composition(adjustedIndex(biK, length), r, items);
   }
 
   /// Returns the index of `selection` in the list of arranged Compositions.
@@ -37,7 +52,7 @@ class Compositions<T> extends Combinatorics<T> {
   BigInt indexOf(List<T> selection, [BigInt? start]) {
     start = start ?? BigInt.zero;
     if (contains(selection)) {
-      BigInt result = _inverseComposition(selection, _items);
+      BigInt result = inverseComposition(selection, _items);
       if (result >= start) {
         return result;
       } else {
@@ -50,9 +65,9 @@ class Compositions<T> extends Combinatorics<T> {
 
   /// Returns whether `x` is in the pseudo-list.
   @override
-  bool contains(List<T> x) => _itemsExistInUniversal(x, _items);
+  bool contains(List<T> x) => itemsExistInUniversal(x, _items);
 
   @override
   String toString() =>
-      'Pseudo-list containing all $length $r-Compositions of items from $items.';
+      "Pseudo-list containing all $length $r-Compositions of items from $items.";
 }