+++ lib/src/compounds.dart
@@ -1,32 +1,48 @@
-part of trotter;
+import "combinatorics.dart" show Combinatorics;
+import "helpers.dart"
+    show
+        nPr,
+        itemsAreUnique,
+        indexFromIntOrBigInt,
+        compound,
+        adjustedIndex,
+        inverseCompound,
+        itemsExistInUniversal;
 
-/// A pseudolist of compounds.
+/// A pseudo-list of compounds.
 ///
-/// A pseudo-list 'containing' all the compounds (permutations of
+/// A pseudo-list "containing" all the compounds (permutations of
 /// unspecified length) of objects taken from the list `elements`.
 ///
 /// _Example_
 ///
-///     final com = Compounds(letters('abcd'));
-///     print('There are ${com.length} subsets of the objects');
-///     print('in ${com.items}.');
-///     print('The first subset is ${com[0]}.');
+///     final com = Compounds(letters("abcd"));
+///     print("There are ${com.length} subsets of the objects");
+///     print("in ${com.items}.");
+///     print("The first subset is ${com[0]}.");
 ///
 
 class Compounds<T> extends Combinatorics<T> {
-  Compounds(List<T> items) {
-    if (!_itemsAreUnique(items)) throw Exception('Items are not unique.');
-
-    _items = List<T>.from(items);
-    _length =
-        (List<BigInt>.generate(items.length + 1, (r) => _nPr(items.length, r))
-            .fold<BigInt>(BigInt.zero, (a, b) => a + b));
+  Compounds(List<T> items)
+      : _items = List<T>.from(items, growable: false),
+        _length = (List<BigInt>.generate(
+                items.length + 1, (r) => nPr(items.length, r))
+            .fold<BigInt>(BigInt.zero, (a, b) => a + b)) {
+    if (!itemsAreUnique(items)) throw Exception("Items are not unique.");
   }
+
+  final List<T> _items;
+  @override
+  List<T> get items => List<T>.from(_items, growable: false);
+
+  final BigInt _length;
+  @override
+  BigInt get length => _length;
 
   @override
   List<T> operator [](Object k) {
-    BigInt biK = _indexFromIntOrBigInt(k);
-    return _compound(_adjustedIndex(biK, length), items);
+    BigInt biK = indexFromIntOrBigInt(k);
+    return compound(adjustedIndex(biK, length), items);
   }
 
   /// Returns the index of `subset` in the list of arranged subsets.
@@ -34,7 +50,7 @@ class Compounds<T> extends Combinatorics<T> {
   BigInt indexOf(List<T> compound, [BigInt? start]) {
     start = start ?? BigInt.zero;
     if (contains(compound)) {
-      BigInt result = _inverseCompound(compound, _items);
+      BigInt result = inverseCompound(compound, _items);
       if (result >= start) {
         return result;
       } else {
@@ -48,9 +64,9 @@ class Compounds<T> extends Combinatorics<T> {
   /// Returns whether `x` is in the pseudo-list.
   @override
   bool contains(List<T> x) =>
-      _itemsExistInUniversal(x, _items) && _itemsAreUnique(x);
+      itemsExistInUniversal(x, _items) && itemsAreUnique(x);
 
   @override
   String toString() =>
-      'Pseudo-list containing all $length compounds of items from $items.';
+      "Pseudo-list containing all $length compounds of items from $items.";
 }