+++ lib/src/helpers.dart
@@ -1,34 +1,32 @@
-part of trotter;
-
 /// Cache to store calculated factorials.
-Map<int, BigInt> _factCache = {};
+final Map<int, BigInt> factCache = {};
 
 /// Calculates `n`!
-BigInt _fact(int n) => _factCache.containsKey(n)
-    ? _factCache[n]!
-    : (n < 2 ? BigInt.one : _factCache[n] = BigInt.from(n) * _fact(n - 1));
+BigInt factorial(int n) => factCache.containsKey(n)
+    ? factCache[n]!
+    : (n < 2 ? BigInt.one : factCache[n] = BigInt.from(n) * factorial(n - 1));
 
 /// Calculates the number of permutations of `r` items taken from `n`.
-BigInt _nPr(int n, int r) => _fact(n) ~/ _fact(n - r);
+BigInt nPr(int n, int r) => factorial(n) ~/ factorial(n - r);
 
 /// Calculates the number of combinations of `r` items taken from `n`.
-BigInt _nCr(int n, int r) => _nPr(n, r) ~/ _fact(r);
+BigInt nCr(int n, int r) => nPr(n, r) ~/ factorial(r);
 
 /// Returns the items in `arrangement` in the same order as they appear in `items`.
-List<T> _sortedArrangement<T>(List<T> arrangement, List<T> items) =>
+List<T> sortedArrangement<T>(List<T> arrangement, List<T> items) =>
     (List<T>.from(arrangement))
       ..sort((x, y) => items.indexOf(x).compareTo(items.indexOf(y)));
 
 /// Checks whether the items in `items` are unique.
-bool _itemsAreUnique<T>(List<T> items) => items.toSet().length == items.length;
+bool itemsAreUnique<T>(List<T> items) => items.toSet().length == items.length;
 
 /// Checks whether all items in `items` are in `universal`.
-bool _itemsExistInUniversal<T>(List<T> items, List<T> universal) =>
+bool itemsExistInUniversal<T>(List<T> items, List<T> universal) =>
     items.every((item) => universal.contains(item));
 
 /// Gives `k`th permutation in the ordered list of permutations of
 /// items taken from `items`.
-List<T> _permutationWorker<T>(BigInt k, List<T> items) {
+List<T> permutationWorker<T>(BigInt k, List<T> items) {
   int n = items.length;
   if (n <= 1) {
     return items;
@@ -39,14 +37,14 @@ List<T> _permutationWorker<T>(BigInt k, List<T> items) {
         position =
             group % BigInt.two == BigInt.zero ? biN - mod - BigInt.one : mod;
 
-    return _permutationWorker<T>(group, items.sublist(0, n - 1))
+    return permutationWorker<T>(group, items.sublist(0, n - 1))
       ..insert(position.toInt(), items[n - 1]);
   }
 }
 
 /// Gives the index of `permutation` in the ordered list of permutations of
 /// items taken from `items`.
-BigInt _inversePermutationWorker<T>(List<T> permutation, List<T> items) {
+BigInt inversePermutationWorker<T>(List<T> permutation, List<T> items) {
   if (permutation.length == 1) {
     return BigInt.zero;
   }
@@ -54,7 +52,7 @@ BigInt _inversePermutationWorker<T>(List<T> permutation, List<T> items) {
 
   BigInt biN = BigInt.from(n),
       index = BigInt.from(permutation.indexOf(items.last)),
-      group = _inversePermutationWorker<T>(
+      group = inversePermutationWorker<T>(
           permutation.where((x) => x != items.last).toList(),
           items.sublist(0, items.length - 1));
   return biN * group +
@@ -63,25 +61,25 @@ BigInt _inversePermutationWorker<T>(List<T> permutation, List<T> items) {
 
 /// Gives `k`th combination in the ordered list of combinations of
 /// `r` items taken from `items`.
-List<T> _combination<T>(BigInt k, int r, List<T> items) {
+List<T> combination<T>(BigInt k, int r, List<T> items) {
   if (r == 0) {
     return [];
   }
   int n = items.length, position = 0;
-  BigInt d = _nCr(n - position - 1, r - 1);
+  BigInt d = nCr(n - position - 1, r - 1);
 
   while (k >= d) {
     k -= d;
     position += 1;
-    d = _nCr(n - position - 1, r - 1);
+    d = nCr(n - position - 1, r - 1);
   }
   final tail = items.sublist(position + 1);
-  return [items[position]]..addAll(_combination<T>(k, r - 1, tail));
+  return [items[position]]..addAll(combination<T>(k, r - 1, tail));
 }
 
 /// Gives the index of `combination` in the ordered list of combinations of
 /// items taken from `items`.
-BigInt _inverseCombination<T>(List<T> combination, List<T> items) {
+BigInt inverseCombination<T>(List<T> combination, List<T> items) {
   BigInt helper(List<T> combination, List<T> items) {
     if (combination.isEmpty) {
       return BigInt.zero;
@@ -89,36 +87,36 @@ BigInt _inverseCombination<T>(List<T> combination, List<T> items) {
     int r = combination.length, n = items.length, itemIndex = 0;
     BigInt k = BigInt.zero;
     while (combination[0] != items[itemIndex]) {
-      k += _nCr(n - itemIndex - 1, r - 1);
+      k += nCr(n - itemIndex - 1, r - 1);
       itemIndex += 1;
     }
     return k + helper(combination.sublist(1), items.sublist(itemIndex + 1));
   }
 
-  return helper(_sortedArrangement<T>(combination, items), items);
+  return helper(sortedArrangement<T>(combination, items), items);
 }
 
 /// Gives `k`th composition in the ordered list of compositions of
 /// `r` items taken from `items`.
-List<T> _composition<T>(BigInt k, int r, List<T> items) {
+List<T> composition<T>(BigInt k, int r, List<T> items) {
   int n = items.length, position = 0;
-  BigInt d = _nCr(n + r - position - 2, r - 1);
+  BigInt d = nCr(n + r - position - 2, r - 1);
   while (k >= d) {
     k -= d;
     position += 1;
-    d = _nCr(n + r - position - 2, r - 1);
+    d = nCr(n + r - position - 2, r - 1);
   }
   if (r == 0) {
     return [];
   } else {
     final tail = items.sublist(position);
-    return [items[position]]..addAll(_composition<T>(k, r - 1, tail));
+    return [items[position]]..addAll(composition<T>(k, r - 1, tail));
   }
 }
 
 /// Gives the index of `composition` in the ordered list of compositions of
 /// items taken from `items`.
-BigInt _inverseComposition<T>(List<T> composition, List<T> items) {
+BigInt inverseComposition<T>(List<T> composition, List<T> items) {
   BigInt helper(List<T> composition, List<T> items) {
     if (composition.isEmpty) {
       return BigInt.zero;
@@ -126,40 +124,40 @@ BigInt _inverseComposition<T>(List<T> composition, List<T> items) {
     int n = items.length, r = composition.length, itemIndex = 0;
     BigInt k = BigInt.zero;
     while (composition[0] != items[itemIndex]) {
-      k += _nCr(n + r - itemIndex - 2, r - 1);
+      k += nCr(n + r - itemIndex - 2, r - 1);
       itemIndex += 1;
     }
 
     return k + helper(composition.sublist(1), items.sublist(itemIndex));
   }
 
-  return helper(_sortedArrangement<T>(composition, items), items);
+  return helper(sortedArrangement<T>(composition, items), items);
 }
 
 /// Gives `k`th permutation in the ordered list of permutations of
 /// `r` items taken from `items`.
-List<T> _permutation<T>(BigInt k, int r, List<T> items) {
-  BigInt f = _fact(r), group = k ~/ f, item = k % f;
-  List<T> comb = _combination<T>(group, r, items);
-  return _permutationWorker<T>(item, comb);
+List<T> permutation<T>(BigInt k, int r, List<T> items) {
+  BigInt f = factorial(r), group = k ~/ f, item = k % f;
+  List<T> comb = combination<T>(group, r, items);
+  return permutationWorker<T>(item, comb);
 }
 
 /// Gives the index of `permutation` in the ordered list of permutations of
 /// items taken from `items`.
-BigInt _inversePermutation<T>(List<T> permutation, List<T> items) {
+BigInt inversePermutation<T>(List<T> permutation, List<T> items) {
   int r = permutation.length;
   if (r == 0) {
     return BigInt.zero;
   }
-  final sortedPermutation = _sortedArrangement<T>(permutation, items);
-  BigInt group = _inverseCombination<T>(sortedPermutation, items);
-  return group * _fact(r) +
-      _inversePermutationWorker<T>(permutation, sortedPermutation);
+  final sortedPermutation = sortedArrangement<T>(permutation, items);
+  BigInt group = inverseCombination<T>(sortedPermutation, items);
+  return group * factorial(r) +
+      inversePermutationWorker<T>(permutation, sortedPermutation);
 }
 
 /// Gives `k`th amalgam in the ordered list of amalgams of
 /// `r` items taken from `items`.
-List<T> _amalgam<T>(BigInt k, int r, List<T> items) {
+List<T> amalgam<T>(BigInt k, int r, List<T> items) {
   return List<T>.generate(r, (i) {
     BigInt p = BigInt.from(items.length).pow(r - i - 1), index = k ~/ p;
     k %= p;
@@ -169,7 +167,7 @@ List<T> _amalgam<T>(BigInt k, int r, List<T> items) {
 
 /// Gives the index of `amalgam` in the ordered list of amalgams of
 /// items taken from `items`.
-BigInt _inverseAmalgam<T>(List<T> amalgam, List<T> items) {
+BigInt inverseAmalgam<T>(List<T> amalgam, List<T> items) {
   int r = amalgam.length;
   final n = BigInt.from(items.length);
   final powers = List<BigInt>.filled(r, BigInt.one, growable: false);
@@ -185,8 +183,8 @@ BigInt _inverseAmalgam<T>(List<T> amalgam, List<T> items) {
 
 /// Gives `k`th subset in the ordered list of subsets of
 /// items taken from `items`.
-List<T> _subset<T>(BigInt k, List<T> items) {
-  k = _adjustedIndex(k, BigInt.one << items.length);
+List<T> subset<T>(BigInt k, List<T> items) {
+  k = adjustedIndex(k, BigInt.one << items.length);
   List<T> r = [];
   for (int i = 0; i < items.length; i++) {
     if (k & (BigInt.one << i) != BigInt.zero) {
@@ -198,7 +196,7 @@ List<T> _subset<T>(BigInt k, List<T> items) {
 
 /// Gives the index of `subset` in the ordered list of subsets of
 /// items taken from `items`.
-BigInt _inverseSubset<T>(List<T> subset, List<T> items) {
+BigInt inverseSubset<T>(List<T> subset, List<T> items) {
   BigInt helper(List<T> subset, List<T> items) {
     BigInt k = BigInt.zero, power = BigInt.one;
     for (int index = 0; index < items.length; index++) {
@@ -210,39 +208,39 @@ BigInt _inverseSubset<T>(List<T> subset, List<T> items) {
     return k;
   }
 
-  return helper(_sortedArrangement<T>(subset.toSet().toList(), items), items);
+  return helper(sortedArrangement<T>(subset.toSet().toList(), items), items);
 }
 
 /// Gives `k`th compound in the ordered list of compounds of
 /// items taken from `items`.
-List<T> _compound<T>(BigInt k, List<T> items) {
+List<T> compound<T>(BigInt k, List<T> items) {
   int n = items.length, r;
   for (r = 0; r < n; r++) {
-    BigInt groupSize = _nPr(n, r);
+    BigInt groupSize = nPr(n, r);
     if (k >= groupSize) {
       k -= groupSize;
     } else {
       break;
     }
   }
-  return _permutation<T>(k, r, items);
+  return permutation<T>(k, r, items);
 }
 
 /// Gives the index of `compound` in the ordered list of compounds of
 /// items taken from `items`.
-BigInt _inverseCompound<T>(List<T> compound, List<T> items) {
+BigInt inverseCompound<T>(List<T> compound, List<T> items) {
   BigInt k =
-      (List<BigInt>.generate(compound.length, (r) => _nPr(items.length, r)))
+      (List<BigInt>.generate(compound.length, (r) => nPr(items.length, r)))
           .fold<BigInt>(BigInt.zero, (a, b) => a + b);
-  k += _inversePermutation<T>(compound, items);
+  k += inversePermutation<T>(compound, items);
   return k;
 }
 
 /// Returns an index in the domain [0, n[.
-BigInt _adjustedIndex(BigInt k, BigInt n) => k % n;
+BigInt adjustedIndex(BigInt k, BigInt n) => k % n;
 
 /// Ensures an index is of type BigInt.
-BigInt _indexFromIntOrBigInt(Object k) {
+BigInt indexFromIntOrBigInt(Object k) {
   BigInt biK;
   if (k is int) {
     biK = BigInt.from(k);
@@ -253,3 +251,9 @@ BigInt _indexFromIntOrBigInt(Object k) {
   }
   return biK;
 }
+
+/// Splits string [x] into its characters.
+List<String> characters(String x) => x.split('');
+
+/// Creates a string from the characters in [x].
+String string(List<String> x) => x.join('');