+++ lib/src/subsets.dart
@@ -1,30 +1,45 @@
-part of trotter;
+import "combinatorics.dart" show Combinatorics;
+import "helpers.dart"
+    show
+        itemsAreUnique,
+        indexFromIntOrBigInt,
+        subset,
+        adjustedIndex,
+        inverseSubset,
+        itemsExistInUniversal;
 
 /// A pseudolist of subsets.
 ///
-/// A pseudo-list 'containing' all the subsets of objects taken from
+/// A pseudo-list "containing" all the subsets of objects taken from
 /// the list `elements`.
 ///
 /// _Example_
 ///
-///     final sub = Subsets(characters('abcd'));
-///     print('There are ${sub.length} subsets of the objects');
-///     print('in ${sub.items}.');
-///     print('The first subset is ${sub[0]}.');
+///     final sub = Subsets(characters("abcd"));
+///     print("There are ${sub.length} subsets of the objects");
+///     print("in ${sub.items}.");
+///     print("The first subset is ${sub[0]}.");
 ///
 
 class Subsets<T> extends Combinatorics<T> {
-  Subsets(List<T> items) {
-    if (!_itemsAreUnique(items)) throw Exception('Items are not unique.');
-
-    _items = List<T>.from(items);
-    _length = BigInt.one << items.length;
+  Subsets(List<T> items)
+      : _items = List<T>.from(items, growable: false),
+        _length = BigInt.one << items.length {
+    if (!itemsAreUnique(items)) throw Exception("Items are not unique.");
   }
+
+  final List<T> _items;
+  @override
+  List<T> get items => List<T>.from(_items, growable: false);
+
+  final BigInt _length;
+  @override
+  BigInt get length => _length;
 
   @override
   List<T> operator [](Object k) {
-    BigInt biK = _indexFromIntOrBigInt(k);
-    return _subset(_adjustedIndex(biK, length), items);
+    BigInt biK = indexFromIntOrBigInt(k);
+    return subset(adjustedIndex(biK, length), items);
   }
 
   /// Returns the index of `subset` in the list of arranged subsets.
@@ -32,7 +47,7 @@ class Subsets<T> extends Combinatorics<T> {
   BigInt indexOf(List<T> subset, [BigInt? start]) {
     start = start ?? BigInt.zero;
     if (contains(subset)) {
-      BigInt result = _inverseSubset(subset, _items);
+      BigInt result = inverseSubset(subset, _items);
       if (result >= start) {
         return result;
       } else {
@@ -46,9 +61,9 @@ class Subsets<T> extends Combinatorics<T> {
   /// Returns whether `x` is in the pseudo-list.
   @override
   bool contains(List<T> x) =>
-      _itemsExistInUniversal(x, _items) && _itemsAreUnique(x);
+      itemsExistInUniversal(x, _items) && itemsAreUnique(x);
 
   @override
   String toString() =>
-      'Pseudo-list containing all $length subsets of items from $items.';
+      "Pseudo-list containing all $length subsets of items from $items.";
 }